pyproject.toml
[project]
name = "celery-yaml"
version = "2.2.0"
description = "Easy configuration for celery app using Yaml file"
authors = [{ name = "Guillaume Gauvrit", email = "guillaume@gauvr.it" }]
readme = "README.md"
requires-python = ">=3.9"
dependencies = ["celery>=5.4.0,<6", "PyYAML>=6.0.0,<7", "envsub>=0.1.15,<1"]
[project.urls]
Homepage = "https://github.com/mardiros/celery-yaml"
Documentation = "https://github.com/mardiros/celery-yaml/blob/main/README.md"
Repository = "https://github.com/mardiros/celery-yaml.git"
Issues = "https://github.com/mardiros/celery-yaml/issues"
Changelog = "https://github.com/mardiros/celery-yaml/blob/main/CHANGELOG.md"

[dependency-groups]
dev = [
    "celery>=5.2.0,<6",
    "coverage>=6.2,<7",
    "mypy>=1.13.0,<2",
    "pytest>=8.3.3,<9",
    "pytest-cov>=5.0.0,<6",
    "pyramid-helloworld>=2.0.0",
    "types-pyyaml>=6.0.12.20240917,<7",
]
pyramid = ["pyramid>=1.9,<3", "plaster-yaml>=2.0,<3"]


[tool.pdm.build]
includes = ["src", "CHANGELOG.md"]
excludes = ["tests"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.mypy]
exclude = ["tests/dummy_packages"]
[[tool.mypy.overrides]]
module = "celery_yaml.*"
disallow_any_generics = true
disallow_untyped_defs = true

[[tool.mypy.overrides]]
module = ["celery.*", "plaster_yaml.*"]
ignore_missing_imports = true

[tool.ruff]
line-length = 88
target-version = "py39"

[tool.ruff.lint]
select = [
    "B",   # bug bear security warning
    "I",   # isort import order
    "F",   # pyflakes
    "UP",  # alter when better syntax is available
    "RUF", #  the ruff devleoper's own rules
]

[tool.coverage.report]
exclude_lines = [
    "if TYPE_CHECKING:",
    "except ImportError:",
    "except pkg_resources.DistributionNotFound:",
    "\\s+\\.\\.\\.$",
    "# coverage-ignore",
]

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"
